requirements.txt
streamlit
streamlit-option-menu
torch
torchvision
pillow